Add new project GUID mapping to solution file
A new entry was added to the solution file, specifically a GUID mapping for a project or component. The added line is: `{8F4209CB-FE3E-497A-AD47-D44819EE02A2} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}`. This indicates that a new project or component is now associated with the same solution as the others listed.